+++ packages/core/src/validate/index.ts
@@ -1,5 +1,5 @@
 export { validate, ValidationRunData, ContractValidation } from './run';
-export { ValidationOptions, withValidationDefaults, silenceWarnings } from './overrides';
+export { ValidationOptions, withValidationDefaults } from './overrides';
 export { ValidationErrors } from './error';
 export { RunValidation, ValidationLog, Validation, ValidationResult } from './compat';
 export { ValidationData, ValidationDataCurrent, isCurrentValidationData, concatRunData } from './data';
@@ -12,3 +12,6 @@ export {
   getErrors,
   isUpgradeSafe,
 } from './query';
+
+// Backwards compatibility
+export { silenceWarnings } from '../utils/log';